fix: moving files due to changed file structures
