+++ WEB(FE)/ar/src/components/base/hooks/useHeader.js
@@ -1,6 +1,6 @@
 import { createStoreHook, useDispatch, useSelector, } from 'react-redux';
 import { rootReducer } from '../../../app/store';
-import { UserActions } from '../../../app/UserSlice';
+import { UserActions } from '../../../app/slice/UserSlice';
 
 export default function useHeader() {
   const dispatch = useDispatch();
